requirements.txt
torch>=2.4.0
torchvision>=0.19.0
torchaudio>=2.4.0
accelerate>=0.29.0

transformers>=4.40.0
sentence-transformers>=2.7.0
bitsandbytes>=0.43.0

langchain>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.1
langchain-chroma>=0.1.0
langchain-text-splitters>=0.2.0

chromadb>=0.5.0
flashrank>=0.2.0

pypdf>=4.2.0
python-dotenv>=1.0.0
ebooklib>=0.18
lxml>=5.0.0
beautifulsoup4>=4.12.0

streamlit>=1.35.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
tqdm>=4.66.0